Move application to other package Remove unused usage of moshi dependency Add image and Readme
