+++ app/src/debug/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <application android:name=".App"
+    <application android:name=".ui.App"
         tools:ignore="MissingApplicationIcon">
         <activity
             android:name=".MainActivityTest"
